requirements.txt
Flask==2.3.2
newspaper3k==0.2.8
nltk==3.6.7
pyngrok==5.2.1
langchain_groq
